add service cli
