--- consumer/settings.toml
@@ -1,3 +1,0 @@
-[default]
-DATABASE_URL = "sqlite:///:memory:"
-SEARCH_QUERY= 'topic:"flask-extension flask-extensions"'
