+++ service/pyproject.toml
@@ -1,8 +1,11 @@
 [tool.poetry]
-name = "consumer"
+name = "fexservice"
 version = "0.1.0"
 description = ""
 authors = ["Bruno Rocha <rochacbruno@gmail.com>"]
+
+[tool.poetry.scripts]
+fexservice = "fexservice.cli:cli"
 
 [tool.poetry.dependencies]
 python = "^3.8"
@@ -10,6 +13,7 @@ dynaconf = "^2.2.3"
 pygithub = "^1.47"
 sqlalchemy = "^1.3.15"
 dataset = "^1.3.0"
+typer = {version = "^0.1.0", extras = ["all"]}
 
 [tool.poetry.dev-dependencies]
 pytest = "^4.6"
