+++ service/tests/test_service.py
@@ -1,0 +1,5 @@
+from service import __version__
+
+
+def test_version():
+    assert __version__ == "0.1.0"
